requirements.txt
aiofiles==24.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.17.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.12
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
hashids==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
maxapi==0.9.13
multidict==6.7.1
openai==2.6.1
openpyxl==3.1.5
propcache==0.4.1
puremagic==1.30
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-multipart==0.0.21
PyYAML==6.0.3
rewire==0.10.4
rewire-sqlmodel==0.7.1
rewire_fastapi==0.2.0
sniffio==1.3.1
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
sqlmodel==0.0.27
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.35.0
win32_setctime==1.2.0
yarl==1.22.0
